Move most of the code into the Python package
